+++ example/web/index.dart
@@ -1,11 +1,12 @@
 // import 'package:aspen_assets/aspen_assets.dart' as aspen;
 // import 'dart:html';
 
+import 'package:aspen_assets/aspen_assets.dart';
 import 'package:aspen_playground/assets.dart';
 import 'package:aspen_web/aspen_web.dart';
 
 void main() {
-  JsAsset(text: 'console.log("loading...")')
+  JsAsset(AssetData('testasset', 'console.log("loading...")'))
       .evalAsync()
       .then((_) => print('evalAsync done'));
   print('after evalAsync');